add bulid deps vocab file
