--- dataset/medical3000/graph_vocab.txt
@@ -1,128 +1,0 @@
-<EMPTY>
-<UNK>
-mPunc
-Exp
-Desc
-eCoo
-Root
-mConj
-mAux
-mMod
-Cont
-mPrep
-eSucc
-Host
-Loc
-Poss
-Quan
-mRang
-Time
-Pat
-Clas
-mDegr
-dCont
-Belg
-mTime
-Mann
-Sco
-Datv
-Agt
-dDesc
-dExp
-mNeg
-Cons
-eSelt
-Qp
-eResu
-mMaj
-eAdvt
-mFreq
-eProg
-dClas
-Stat
-Prod
-dTime
-Reas
-Freq
-eRect
-eCau
-Nmod
-eInf
-Dir
-Nfin
-Nini
-Orig
-ePurp
-Seq
-Comp
-Tool
-mVain
-Tdur
-dCons
-Accd
-dDatv
-dBelg
-rExp
-mTone
-Lthru
-Lfin
-Aft
-rDesc
-eCond
-eSupp
-rCont
-Tini
-rPat
-eEqu
-Tfin
-dReas
-Matl
-Sproc
-Sfin
-rCons
-Int
-Trang
-dPoss
-mDir
-mRept
-mPars
-eMetd
-dOrig
-rAgt
-ePrec
-Lini
-rPoss
-rReas
-eConc
-dPat
-dHost
-dAgt
-Tmod
-rMann
-Sini
-rProd
-rBelg
-TDur
-dStat
-dMann
-eSum
-dSco
-dLoc
-rTime
-dProd
-rDatv
-ePref
-dInt
-dAccd
-dComp
-rOrig
-rInt
-rTool
-dSproc
-rDir
-Nvar
-dTrang
-mSepa
-dTfin
-rQuan
-eAban
