--- dataset/sem16/graph_vocab.txt
@@ -1,163 +1,0 @@
-<EMPTY>
-<UNK>
-mPunc
-Desc
-mAux
-Root
-Exp
-Agt
-mPrep
-mMod
-eSucc
-mTime
-eCoo
-Quan
-mConj
-Cont
-Qp
-Loc
-Pat
-mRang
-Host
-Datv
-dCont
-Sco
-Nmod
-Time
-Clas
-mDegr
-Mann
-mNeg
-Poss
-Aft
-mTone
-eResu
-eEqu
-Belg
-dDesc
-Prod
-ePurp
-mFreq
-mDir
-eAdvt
-eProg
-rExp
-Lfin
-Dir
-dExp
-Seq
-dTime
-mVain
-eCau
-Tmod
-rAgt
-Tdur
-mPars
-rCont
-mMaj
-Orig
-dClas
-Reas
-rPat
-eSupp
-Comp
-eSelt
-Tool
-Cons
-Accd
-dMann
-TDur
-Lini
-Freq
-eConc
-rProd
-rTime
-Stat
-rPoss
-eCond
-Matl
-eRect
-rLoc
-Lthru
-eInf
-Tfin
-Tini
-dDatv
-rDatv
-Nvar
-Sproc
-dReas
-Nfin
-Sini
-mSepa
-Sfin
-dPat
-dCons
-rAft
-ePrec
-dLoc
-rBelg
-rMann
-eSum
-rReas
-dOrig
-dStat
-rTool
-dAccd
-dTini
-Nini
-dPoss
-Trang
-eAban
-dComp
-dSproc
-rTdur
-Int
-dBelg
-rCons
-rClas
-rDir
-dSini
-dDir
-ePref
-dHost
-rQuan
-rInt
-dInt
-dSco
-dSfin
-rSco
-eMetd
-dProd
-dTdur
-rDesc
-rAccd
-rTDur
-dTfin
-rLthru
-rMatl
-mRept
-dTDur
-rOrig
-mQuaf
-dFreq
-rSproc
-rLfin
-dNfin
-rHost
-dNini
-dTool
-rQp
-dAgt
-dLfin
-dMatl
-rTrang
-dTrang
-rSfin
-rSini
-rComp
-rTfin
-rLini
-rStat
-dAft
-rNfin
