--- dataset/test.conllu
@@ -1,17 +1,0 @@
-1	妈妈	妈妈	NN	NN	_	5	Agt	5:Agt	_
-2	把	把	P	P	_	4	mPrep	4:mPrep	_
-3	旧	旧	JJ	JJ	_	4	Desc	4:Desc	_
-4	窗帘	窗帘	NN	NN	_	5	Pat	5:Pat	_
-5	撕成	撕成	VV	VV	_	0	Root	0:Root	_
-6	了	了	AS	AS	_	5	mTime	5:mTime	_
-7	抹布	抹布	NN	NN	_	5	Clas	5:Clas	_
-8	。	。	PU	PU	_	5	mPunc	5:mPunc	_
-
-1	我	我	PN	PN	_	_	_	_	_
-2	爱	爱	VV	VV	_	_	_	_	_
-3	他	他	PN	PN	_	_	_	_	_
-
-1	我	我	PN	PN	_	8	Agt	4:Agt	_
-2	一定	一定	AD	AD	_	8	mMod	4:mMod	_
-3	在	在	P	P	_	4	mPrep	4:mPrep	_
-4	九点	九点	NT	NT	_	8	Tfin	0:Root	_
