--- utils/input_utils/__init__.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
--------------------------------------------------
-   File Name：     __init__.py
-   Description :
-   Author :       Liangs
-   date：          2019/7/28
--------------------------------------------------
-   Change Activity:
-                   2019/7/28:
--------------------------------------------------
-"""
