--- utils/input_utils/bertology/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# Created by li huayong on 2019/9/24
-
-
-if __name__ == '__main__':
-    pass
