+++ utils/model/initialization.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+# Created by li huayong on 2019/11/3
+import torch
+import torch.nn as nn
+from fairseq.modules import LayerNorm, MultiheadAttention
+
+
+def init_bert_params(module):
+    """
+    Initialize the weights specific to the BERT Model.
+    This overrides the default initializations depending on the specified arguments.
+        1. If normal_init_linear_weights is set then weights of linear
+           layer will be initialized using the normal distribution and
+           bais will be set to the specified value.
+        2. If normal_init_embed_weights is set then weights of embedding
+           layer will be initialized using the normal distribution.
+        3. If normal_init_proj_weights is set then weights of
+           in_project_weight for MultiHeadAttention initialized using
+           the normal distribution (to be validated).
+    """
+
+    if isinstance(module, nn.Linear):
+        module.weight.data.normal_(mean=0.0, std=0.02)
+        if module.bias is not None:
+            module.bias.data.zero_()
+    if isinstance(module, nn.Embedding):
+        module.weight.data.normal_(mean=0.0, std=0.02)
+    if isinstance(module, MultiheadAttention):
+        module.in_proj_weight.data.normal_(mean=0.0, std=0.02)
+
+
+if __name__ == '__main__':
+    pass
