--- utils/model_utils/__init__.py
@@ -1,14 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
--------------------------------------------------
-   File Name：     __init__.py
-   Description :
-   Author :       Liangs
-   date：          2019/7/28
--------------------------------------------------
-   Change Activity:
-                   2019/7/28:
--------------------------------------------------
-"""
-from .sort import unsort, sort, tensor_unsort
-from .parser_funs import sdp_decoder, parse_semgraph
